Impl all asset precompile interfaces

+++ runtime/darwinia/src/pallets/assets.rs
@@ -23,7 +23,7 @@ impl pallet_assets::Config for Runtime {
 	type ApprovalDeposit = ConstU128<0>;
 	type AssetAccountDeposit = ConstU128<2>;
 	type AssetDeposit = ConstU128<2>;
-	type AssetId = u32;
+	type AssetId = u64;
 	type Balance = u128;
 	type Currency = Balances;
 	type Extra = ();
